Get the last fetched scrobble date from the DB instead
If you start listening to a track at 23:59:59 UTC and fetch scrobbles at 00:00
UTC the next day, it won't fetch that track for the previous day (as it's
now-playing), and will proceed to fetch the current day. `recenttracks` will be
empty, but it will still save the JSON, which affects `get_last_fetched_date`
during the next run. That results in the previously now-playing track never
being fetched.

Instead, change `get_last_fetched_date` to query the DB instead.